+++ asgard/clients/chronos/errors.py
@@ -1,0 +1,2 @@
+class Htp404(Exception):
+    pass
